+++ packages/tools/pie-wrapper-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @justeattakeaway/pie-wrapper-react
 
+## 0.9.1
+
+### Patch Changes
+
+- [Added] - New ActionProps type for modal leading/supporting actions ([#670](https://github.com/justeattakeaway/pie/pull/670)) by [@xander-marjoram](https://github.com/xander-marjoram)
+
 ## 0.9.0
 
 ### Minor Changes